move tool support to a part file
